get_erc20_tokens/requirements.txt
web3
requests
pandas
python-dotenv
google-cloud-bigquery
